Pass state to plugin init and fix image path
Path to save images to was all kinds of broken.

The plugin load interface is still a bit wonky - state and plugin confs
are passed around, but plugin confs is just a map out of config. Maybe
in the future config will be in the state, and plugin-loader can just
reach into there and get the plugin confs.