Fix plugin-loader and avoid eval
Eval was producing very strange errors realated to expanding literals -
it had never been working, but there was an illusion when plug conf and
state data structures looked a particular, lucky way. I think this was
due to putting them at the tail of a list.

This way is far more elegant anyway, and is what I would have been using
all along if I had known about it.