Change async-dl loop to future
It was spawned with clojure.core.async/thread. Also, make sure we end
the thread when we are done by closing the channel. This is a bit saner
than just killing the thread because it makes sure we process any items
in the queue before we exit.

For some reason, there was no need to end the coroutine using the old
method - probably some undocumented feature of the thread pools used by
the async system.