Pull config out of state
Package them both in a map called "context". This also gives core
components like async-dl and chan-th the same view as plugin-loader and
plugins. This is much preferable to idiosyncratic combinations of state,
conf, plugins conf, conf of just one plugin, etc. being passed around.

However, this is not a complete overhaul, and the state ns still exists
for now.