Remove all config merging
Replace all config merging with placing configs in a sequence. This
should have the same effect in all cases, but will preserve them in a
list, instead of flattenning them into one map. This paves the way for,
for example.  putting configs behind atoms to allow mutation. In
general, this is the completion of the migration to conf-get.