Allow -d and a command to be used simultaneously
Previously, if -d was passed the command was silently ignored.

This requires async-dl/deinit improvements so that the shutdown can be
prompt. Without the async-dl/deinit interlock, shutdown-agents could
potentially kill the async-dl thread before it is done, and without
shutdown-agents, we have to wait a full minute for the process to end.
This wait is a limitation of the clojure runtime, and has to do with
ending worker threads.