Check for commands that cannot parse
Previously, was forgetting to check the error altogether. I
(misremembered) the parse API as returning empty :arguments when there
was an error, but it still returns the positional arguments when all
that it could not parse was a flag.

The error checking in core was correct. This checking in core should
probably be removed in favor of this checking.