Don't quit when failing to init thread
Previously, quit right away if there was any kind of network response we
did not like while initing a thread. For example, if an ID was mistyped,
this would exit the entire program.

There was no reason to do this, and it is trivial to change to just
logging the error and continuing. The only non-obvious change is
changing the code in the future to handle threads failing to init -
previously we would still try to refresh one, even if the init failed.

I additionally removed the method, since it was all called in one place,
and would have had to be renamed anyway.