Fix extra nesting and post ordering bugs
Before this, every time a thread was updated the new processed posts
were added using conj. However, since the old posts value was some kind
of sequence (list? iter?) the result was something like:
[[new-post-1 new-post-2] old-post-1 old-post-2]
In other words, a list of new-posts was prepended to the old posts. This
not only was the wrong order, but it was nesting an extra level of
collection.

Being careful about the type and the append fixes both at once, though
may have performance downsides for threads with many posts.