Only store threads that properly initted
Only keep thread handles of threads that actually init. This should save
some resources by garbage collecting the futures that don't init.

This also paves the way for threads only being added if they are not
already present. If threads which failed to init were kept in the map,
they could never be re-added, even if that re-add would result in a
proper init.