Separate init from processing current posts
Previously, init-thread downloaded the thread and processed the current
posts, all in one. Combined with thread-manager, this meant that there
was no way to init the thread (thus doing error checking, eg. ensuring
the thread exists) without also processing the posts.

This separation also fixes the issue where the thread executing
thread-manger code (afaik the main thread) was blocked while a thread
was doing the first init (meaning it was blocked until all the images
present were downloaded). This would subtly delay handling other thread
inits. The other thread inits were queued in the yet-to-be-read socket
buffer - NOT a principled queueing strategy.