Export thread every time it changes
- Reuse the core export logic from the old implementation.
- Add test for trim-for-export
- Test the actual exported text by mocking the file with a StringWriter.

Note that this does not export in the same format as the old
implementation, because it leaves the map of posts rather than making
them back into a sequence.

Additionally, there are potential performance issues with writing every
time the agent value changes. For example, if many images are being
downloaded and each is changing the thread, many writes will happen when
only one or two are meaningful.