Fix update of posts
- Fix handling of posts whose JSON has changed.
- Excise clojure.data/diff, which was a bit worrying from the get-go.
- Slight test data changes to more faciliate more realistic update test.
- Modify `refresh-integrate-new-image-test` to test this case.
- Close unrelated test hole, use sorted map in `trim-for-export-test`

Because `clojure.data/diff` was used to find new posts, when the JSON
contents of a post changed it would show up as a skeleton of a post in
the new-only collection (ie. only the sparse tree that changed).
Because the "no" field containing the ID was missing from `new-only`, it
would end up with key `nil` in the map.

This was exposed so late because it requires:
- Post JSON to be mutable (this is true of the OP only of threads which
  are actively changing, ie. hard to see in slow moving threads) AND
- The post data structure to be examined closely or used (made visible
  by JSON export).

The bug was exposed when the thread was processed for export because the
json renderer rejected `nil` map keys, but was present since the
beginning of the rewrite.