Safer but slower limited-downloader daemon stop
The interrupt can cause exceptions when the thread is interrupted in
certain places. This manifested as errors during test shutdown when it
was interrupted during a `deref`.

Removing the interrupt means the thread keeps running at least until the
current sleep ends, which should be very short, since the sleeps are
within the timing loop. Additionally, the thread is marked as a daemon
thread so it should not keep the JVM from exiting.