Reimplement verifying dir before create
New implementation of the old feature, for the new thread system. This
time error messages are easier to work with (rather than being logged
directly).

thread-manager makes the create-or-not call, but chan-th has all of the
verification logic. This means that when threads get created is easy to
see and understand, and all the error logs are in on place. This will
also allow other verifications to be added in one place in chan-th with
little to no change in thread-manager.