Do not add threads overzealously
Don't add until config has minimal keys: "board", "thread", and "dir"
all must be present before it is given to thread-manager. This allows a
config to be built up in pieces.

Ignore things that are not plain files: this mainly ignores directories,
but will also ignore what each platform considers "special files".

Check if thread-manager has the thread before adding it: This paves the
way for reconfiguration, and is necessary for sane mocking to keep the
test passing.

Some refactoring of passing files vs. paths is necessary to use the file
APIs for these filters.