Change config seq into association seq
Change from sequence of conf to sequence of [name conf] tuple.  The keys
encode the source of the config, and should be unique.

This will be used for a more proper reconfigure implementation in the
future.

- Remove conf-seq, since it provides little with this new format.
- Add `base` so that tests can quickly make a map into a config seq.
- Minimal test retrofits

This contains one major wart, which is that conf-get still supports
plain maps as configs. This is to keep tests working with minimal
modifications, and will be fixed in the future.