Eliminate support for raw map config
Avoid overly complex type variation in config (ie "is it a map this
time, or an association seq"). This allows all users to treat it
consistently, and somewhat simplifies conf-get.

However, the main motivation is still cleaner reconfig of threads, which
will depend on configs consistently taking this form.

This also cleans up test iffiness remaining after the previous change.