Core thread description need not be in file
Previously, the file must directly include "board", "thread", and "dir".
With this chance, some or all of those attributes can be inherited from
previous configs, such as the config on the command line when starting
the program, or the configuration file.

This makes it possible to produce shorter per-thread config files.
Additionally, this paves the way for some simplification of
thread-manager (or consolidation of logic into thread mananger) when
remote-control is removed.

- Refactor configuration checking logic in config-reload
- Change API between config-reload and thread-manager to make supplying
  complete config possible.
- Extensive test changes to test that new configuration takes effect.